% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main_sigGeneSNPPairs.R
\name{callSigGeneSNPPairs}
\alias{callSigGeneSNPPairs}
\title{callSigGeneSNPPairs}
\usage{
callSigGeneSNPPairs(
  exprFile,
  covFile,
  genotypeFile,
  tabixProgram,
  outputDir,
  approach,
  cisDistance = 1e+06,
  MAFThreshold = 0.01,
  MASamplesThreshold = 10,
  numOfCores = 1,
  FDR_eGene = 0.05,
  sigGeneSNPPairMethod = NULL,
  percent = 1
)
}
\arguments{
\item{exprFile}{Must be the same as \code{exprFile} in \code{ClipperQTL()}.}

\item{covFile}{Must be the same as \code{covFile} in \code{ClipperQTL()}.}

\item{genotypeFile}{Must be the same as \code{genotypeFile} in \code{ClipperQTL()}.}

\item{tabixProgram}{Must be the same as \code{tabixProgram} in \code{ClipperQTL()}.}

\item{outputDir}{Must be the same as \code{outputDir} in \code{ClipperQTL()}. Must be recognizable in C++. "~" may not be recognized as the home directory in C++; the user may need to spell out the directory name instead.}

\item{approach}{Must be the same as \code{approach} in \code{ClipperQTL()} (\code{"standard"} or \code{"Clipper"}). Helps determine \code{sigGeneSNPPairMethod}.}

\item{cisDistance}{Must be the same as \code{cisDistance} in \code{ClipperQTL()}.}

\item{MAFThreshold}{Must be the same as \code{MAFThreshold} in \code{ClipperQTL()}.}

\item{MASamplesThreshold}{Must be the same as \code{MASamplesThreshold} in \code{ClipperQTL()}.}

\item{numOfCores}{The number of cores to be used. Default is \code{1}.}

\item{FDR_eGene}{The target FDR for eGene identification. Default is \code{0.05}.}

\item{sigGeneSNPPairMethod}{Key argument. Must be \code{"FastQTL"}, \code{"topPercent"}, or \code{NULL}. See below.}

\item{percent}{Key argument. Only relevant if \code{"topPercent"} is used. Default is \code{1}, which means the top 1\% local common SNPs (in terms of significance of association) will be identified as significant for each eGene. See below.}
}
\description{
This function is used for identifying significant gene-SNP pairs.
}
\details{
Most arguments in this function (from \code{exprFile} to \code{MASamplesThreshold}) must be the same as those used in \code{ClipperQTL()}. However, since \code{outputDir} will be used in C++ in this function, it must be recognizable in C++ (see above).

The main method parameters of \code{callSigGeneSNPPairs()} are \code{sigGeneSNPPairMethod} and \code{percent}. \code{sigGeneSNPPairMethod} must be \code{"FastQTL"}, \code{"topPercent"}, or \code{NULL}. \code{percent} is only relevant if \code{"topPercent"} is used.

If \code{sigGeneSNPPairMethod="FastQTL"}, then significant gene-SNP pairs will be identified using the method in FastQTL (see Algorithm S2 of reference; inverse functions of cumulative distribution functions are replaced by quantile functions).

If \code{sigGeneSNPPairMethod="topPercent"}, then the top 1\% local common SNPs (in terms of significance of association) will be identified as significant for each eGene (if \code{percent=1}, that is).

If \code{sigGeneSNPPairMethod=NULL}, then if \code{approach="standard"}, \code{sigGeneSNPPairMethod} will be set to \code{"FastQTL"}; if \code{approach="Clipper"}, \code{sigGeneSNPPairMethod} will be set to \code{"topPercent"}.

If \code{approach="standard"}, then \code{sigGeneSNPPairMethod} can be either \code{"FastQTL"} or \code{"topPercent"}. If \code{approach="Clipper"}, then \code{sigGeneSNPPairMethod} cannot be \code{"FastQTL"}.

\code{callSigGeneSNPPairs()} outputs several files in the output directory. The most important one is named "_resultSigGeneSNPPairs_FastQTL.rds" or "_resultSigGeneSNPPairs_topPercent.rds" (depending on \code{sigGeneSNPPairMethod}), which can be read into R with \code{readRDS()}. Each row corresponds to a significant gene-SNP pair.
}
\references{
Heather J. Zhou, Xinzhou Ge, and Jingyi Jessica Li. ClipperQTL: ultrafast and powerful eGene identification method. bioRxiv, 2023.
}
