% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Main.R
\name{ClipperQTL}
\alias{ClipperQTL}
\title{ClipperQTL}
\usage{
ClipperQTL(
  exprFile,
  covFile,
  genotypeFile,
  tabixProgram,
  outputDir,
  approach = NULL,
  B = NULL,
  MAFThreshold = 0.01,
  MASamplesThreshold = 10,
  numOfChunksTarget = 100,
  seed = 1,
  numOfCores = 1
)
}
\description{
ClipperQTL
}
