% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main_sigGeneSNPPairs.R
\name{callSigGeneSNPPairs}
\alias{callSigGeneSNPPairs}
\title{callSigGeneSNPPairs}
\usage{
callSigGeneSNPPairs(
  exprFile,
  covFile,
  genotypeFile,
  tabixProgram,
  outputDir,
  approach,
  cisDistance = 1e+06,
  MAFThreshold = 0.01,
  MASamplesThreshold = 10,
  numOfCores = 1,
  FDR_eGene = 0.05,
  sigGeneSNPPairMethod = NULL,
  percent = 1
)
}
\arguments{
\item{outputDir}{Must be recognizable in C++. "~" may not be recognized as the home directory in C++; spell out the directory name instead.}

\item{sigGeneSNPPairMethod}{If approach is standard, then default is "FastQTL". If approach is Clipper, then default is "topPercent".}

\item{percent}{Between 0 and 100. For example, 1 means the top 1% SNPs for each eGene are outputted.}
}
\description{
callSigGeneSNPPairs
}
