% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main_ClipperQTL.R
\name{ClipperQTL}
\alias{ClipperQTL}
\title{ClipperQTL}
\usage{
ClipperQTL(
  exprFile,
  covFile,
  genotypeFile,
  tabixProgram,
  outputDir,
  approach = "standard",
  B = 1000,
  cisDistance = 1e+06,
  MAFThreshold = 0.01,
  MASamplesThreshold = 10,
  numOfChunksTarget = 100,
  seed = 1,
  numOfCores = 1
)
}
\description{
ClipperQTL
}
